Move to string formatting.
